# EAGLE cache fix for HiCache