<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

#10846

## Accept Length
```
python3 -m sglang.launch_server --model meta-llama/Llama-3.1-8B-Instruct --speculative-algo EAGLE3 \
    --speculative-draft-model-path lmsys/sglang-EAGLE3-LLaMA3.1-Instruct-8B  \
    --speculative-num-steps 2 --speculative-eagle-topk 1 --speculative-num-draft-tokens 3 \
    --dtype float16 --page-size 1 --enable-hierarchical-cache --hicache-ratio 1.2

python3 bench_sglang_eagle.py --parallel 1 --num-questions 10
```
main:
```
#questions: 10, Throughput: 333.37 token/s, Acceptance length: 2.25
```
this PR:
```
#questions: 10, Throughput: 351.13 token/s, Acceptance length: 2.39
```

## Cache Hit Rate
```
 python3 benchmark/hicache/bench_multiturn.py --model-path meta-llama/Llama-3.1-8B-Instruct --disable-random-sample \
--output-length 1 --request-length 2048 \
--num-clients 80 --num-rounds 10 --max-parallel 4 --request-rate 16 \
--ready-queue-policy random --disable-auto-run --enable-round-barrier
```

main (page_size 1):
```
Performance metrics summary:
  Total requests: 800 at 16.0 requests per second
  Average Prompt Length: 11687.11 tokens
  Average Output Length: 1.00 tokens
  Average TTFT: 0.39
  P90 TTFT: 1.00
  Median TTFT: 0.22
  Average latency: 0.39
  P90 latency: 1.01
  Median latency: 0.22
  Input token throughput: 89428.71 tokens per second
  Output token throughput: 7.65 tokens per second
  Request Throughput: 7.65 requests per second
  Cache Hit Rate: 0.680656
Per-round metrics:
  Round 0: Average TTFT = 0.07s, Cache Hit Rate = 0.000461 (80 requests)
  Round 1: Average TTFT = 0.11s, Cache Hit Rate = 0.500241 (80 requests)
  Round 2: Average TTFT = 0.14s, Cache Hit Rate = 0.665864 (80 requests)
  Round 3: Average TTFT = 0.17s, Cache Hit Rate = 0.750331 (80 requests)
  Round 4: Average TTFT = 0.20s, Cache Hit Rate = 0.800206 (80 requests)
  Round 5: Average TTFT = 0.25s, Cache Hit Rate = 0.833090 (80 requests)
  Round 6: Average TTFT = 0.32s, Cache Hit Rate = 0.856926 (80 requests)
  Round 7: Average TTFT = 0.54s, Cache Hit Rate = 0.820319 (80 requests)
  Round 8: Average TTFT = 0.81s, Cache Hit Rate = 0.637789 (80 requests)
  Round 9: Average TTFT = 1.25s, Cache Hit Rate = 0.410064 (80 requests)
```

main (page_size 64):
```
Performance metrics summary:
  Total requests: 800 at 16.0 requests per second
  Average Prompt Length: 11732.38 tokens
  Average Output Length: 1.00 tokens
  Average TTFT: 0.37
  P90 TTFT: 0.91
  Median TTFT: 0.19
  Average latency: 0.37
  P90 latency: 0.91
  Median latency: 0.19
  Input token throughput: 93484.75 tokens per second
  Output token throughput: 7.97 tokens per second
  Request Throughput: 7.97 requests per second
  Cache Hit Rate: 0.687901
Per-round metrics:
  Round 0: Average TTFT = 0.07s, Cache Hit Rate = 0.000000 (80 requests)
  Round 1: Average TTFT = 0.10s, Cache Hit Rate = 0.492531 (80 requests)
  Round 2: Average TTFT = 0.13s, Cache Hit Rate = 0.661020 (80 requests)
  Round 3: Average TTFT = 0.16s, Cache Hit Rate = 0.746481 (80 requests)
  Round 4: Average TTFT = 0.19s, Cache Hit Rate = 0.797145 (80 requests)
  Round 5: Average TTFT = 0.21s, Cache Hit Rate = 0.830513 (80 requests)
  Round 6: Average TTFT = 0.29s, Cache Hit Rate = 0.855021 (80 requests)
  Round 7: Average TTFT = 0.55s, Cache Hit Rate = 0.832119 (80 requests)
  Round 8: Average TTFT = 0.83s, Cache Hit Rate = 0.630734 (80 requests)
  Round 9: Average TTFT = 1.17s, Cache Hit Rate = 0.461394 (80 requests)
```

this PR (page_size 1):
```
Performance metrics summary:
  Total requests: 800 at 16.0 requests per second
  Average Prompt Length: 11737.81 tokens
  Average Output Length: 1.00 tokens
  Average TTFT: 0.38
  P90 TTFT: 0.95
  Median TTFT: 0.20
  Average latency: 0.38
  P90 latency: 0.95
  Median latency: 0.20
  Input token throughput: 91844.92 tokens per second
  Output token throughput: 7.82 tokens per second
  Request Throughput: 7.82 requests per second
  Cache Hit Rate: 0.684252
Per-round metrics:
  Round 0: Average TTFT = 0.07s, Cache Hit Rate = 0.000000 (80 requests)
  Round 1: Average TTFT = 0.11s, Cache Hit Rate = 0.500007 (80 requests)
  Round 2: Average TTFT = 0.14s, Cache Hit Rate = 0.665707 (80 requests)
  Round 3: Average TTFT = 0.16s, Cache Hit Rate = 0.750213 (80 requests)
  Round 4: Average TTFT = 0.19s, Cache Hit Rate = 0.800117 (80 requests)
  Round 5: Average TTFT = 0.22s, Cache Hit Rate = 0.833013 (80 requests)
  Round 6: Average TTFT = 0.31s, Cache Hit Rate = 0.856861 (80 requests)
  Round 7: Average TTFT = 0.54s, Cache Hit Rate = 0.820166 (80 requests)
  Round 8: Average TTFT = 0.83s, Cache Hit Rate = 0.628042 (80 requests)
  Round 9: Average TTFT = 1.23s, Cache Hit Rate = 0.442725 (80 requests)
```
this PR (page_size 64):
```
Performance metrics summary:
  Total requests: 800 at 16.0 requests per second
  Average Prompt Length: 11692.25 tokens
  Average Output Length: 1.00 tokens
  Average TTFT: 0.38
  P90 TTFT: 0.93
  Median TTFT: 0.21
  Average latency: 0.38
  P90 latency: 0.93
  Median latency: 0.22
  Input token throughput: 91342.23 tokens per second
  Output token throughput: 7.81 tokens per second
  Request Throughput: 7.81 requests per second
  Cache Hit Rate: 0.684098
Per-round metrics:
  Round 0: Average TTFT = 0.07s, Cache Hit Rate = 0.000000 (80 requests)
  Round 1: Average TTFT = 0.10s, Cache Hit Rate = 0.492344 (80 requests)
  Round 2: Average TTFT = 0.13s, Cache Hit Rate = 0.660773 (80 requests)
  Round 3: Average TTFT = 0.16s, Cache Hit Rate = 0.746779 (80 requests)
  Round 4: Average TTFT = 0.20s, Cache Hit Rate = 0.797447 (80 requests)
  Round 5: Average TTFT = 0.24s, Cache Hit Rate = 0.830512 (80 requests)
  Round 6: Average TTFT = 0.31s, Cache Hit Rate = 0.854866 (80 requests)
  Round 7: Average TTFT = 0.52s, Cache Hit Rate = 0.821278 (80 requests)
  Round 8: Average TTFT = 0.83s, Cache Hit Rate = 0.615397 (80 requests)
  Round 9: Average TTFT = 1.21s, Cache Hit Rate = 0.460686 (80 requests)
```